经销商实体创建

See merge request deepexi/xpaas-backend/commodity-domain/deepexi-channel-center/deepexi-channel-center!15

+++ xpaas-channel-center-channel/src/main/java/com/deepexi/channel/businness/DistributorBusinessService.java
@@ -12,6 +12,6 @@ import java.util.List;
  */
 public interface DistributorBusinessService {
 
-   int create(DistributorDTO dto);
+   long create(DistributorDTO dto);
 
 }
